(BOLT-167) Move PuppetDB::Config into bolt proper
