+++ lib/bolt_ext/puppetdb_inventory.rb
@@ -1,5 +1,6 @@
 #!/usr/bin/env ruby
 
+require 'bolt/puppetdb/config'
 require 'json'
 require 'httpclient'
 require 'optparse'
@@ -67,77 +68,6 @@ module Bolt
       end
     end
 
-    class Config
-      DEFAULT_TOKEN = File.expand_path('~/.puppetlabs/token')
-      DEFAULT_CONFIG = File.expand_path('~/.puppetlabs/client-tools/puppetdb.conf')
-
-      def initialize(config_file, options)
-        @settings = load_config(config_file)
-        @settings.merge!(options)
-
-        expand_paths
-        validate
-      end
-
-      def load_config(filename)
-        if filename
-          if File.exist?(filename)
-            config = JSON.parse(File.read(filename))
-          else
-            raise "config file #{filename} does not exist"
-          end
-        elsif File.exist?(DEFAULT_CONFIG)
-          config = JSON.parse(File.read(DEFAULT_CONFIG))
-        else
-          config = {}
-        end
-        config.fetch('puppetdb', {})
-      end
-
-      def token
-        return @token if @token
-        if @settings['token']
-          File.read(@settings['token'])
-        elsif File.exist?(DEFAULT_TOKEN)
-          File.read(DEFAULT_TOKEN)
-        end
-      end
-
-      def [](key)
-        @settings[key]
-      end
-
-      def expand_paths
-        %w[cacert cert key token].each do |file|
-          @settings[file] = File.expand_path(@settings[file]) if @settings[file]
-        end
-      end
-
-      def validate_file_exists(file)
-        if @settings[file] && !File.exist?(@settings[file])
-          raise "#{file} file #{@settings[file]} does not exist"
-        end
-      end
-
-      def validate
-        unless @settings['server_urls']
-          raise "server_urls must be specified in the config file or with --url"
-        end
-        unless @settings['cacert']
-          raise "cacert must be specified in the config file or with --cacert"
-        end
-
-        if (@settings['cert'] && !@settings['key']) ||
-           (!@settings['cert'] && @settings['key'])
-          raise "cert and key must be specified together"
-        end
-
-        validate_file_exists('cacert')
-        validate_file_exists('cert')
-        validate_file_exists('key')
-      end
-    end
-
     class CLI
       def initialize(args)
         @args = args
@@ -158,7 +88,7 @@ module Bolt
           end
           opts.on('--token-file TOKEN',
                   "Path to the token file",
-                  "Default: #{Config::DEFAULT_TOKEN} if present") do |token|
+                  "Default: #{Bolt::PuppetDB::Config::DEFAULT_TOKEN} if present") do |token|
             @cli_opts['token'] = token
           end
           opts.on('--url URL', "The URL of the PuppetDB server to connect to") do |url|
@@ -166,7 +96,7 @@ module Bolt
           end
           opts.on('--config CONFIG',
                   "The puppetdb.conf file to read configuration from",
-                  "Default: #{Config::DEFAULT_CONFIG} if present") do |file|
+                  "Default: #{Bolt::PuppetDB::Config::DEFAULT_CONFIG} if present") do |file|
             @config_file = File.expand_path(file)
           end
           opts.on('--output FILE', '-o FILE',
@@ -211,7 +141,7 @@ query results.
           raise "Unknown argument(s) #{positional_args.join(', ')}"
         end
 
-        config = Config.new(@config_file, @cli_opts)
+        config = Bolt::PuppetDB::Config.new(@config_file, @cli_opts)
         @puppetdb_client = Client.from_config(config)
 
         unless File.readable?(inventory_file)
